Minor refactoring updates
