+++ test/conftest.py
@@ -1,5 +1,4 @@
 import logging
-import os
 import random
 from pathlib import Path
