+++ test/eks/test_vars.py
@@ -1,9 +1,68 @@
+import uuid
+from pathlib import Path
 from typing import Any
 
+import boto3
 import pytest
 
 from ..utils import MotoServer, Terraform
 from . import Cluster
+
+
+@pytest.fixture(scope="module")
+def eks_client(moto_server: MotoServer) -> ...:
+    return boto3.client("eks", endpoint_url=moto_server.endpoint)
+
+
+@pytest.fixture(scope="module")
+def this_dir() -> Path:
+    return Path(__file__).parent
+
+
+@pytest.fixture(scope="module", autouse=True)
+def vpc(ec2) -> ...:
+    return ec2.create_vpc(CidrBlock="10.0.0.0/16")
+
+
+@pytest.fixture(scope="module", autouse=True)
+def subnets(vpc) -> None:
+    s1 = vpc.create_subnet(
+        AvailabilityZone="eu-west-1a", CidrBlock="10.0.10.0/24"
+    )
+    s2 = vpc.create_subnet(
+        AvailabilityZone="eu-west-1b", CidrBlock="10.0.11.0/24"
+    )
+    return s1, s2
+
+
+@pytest.fixture(scope="module", autouse=True)
+def sg(ec2, vpc) -> None:
+    sg = ec2.create_security_group(
+        GroupName="ssh", Description="ssh", VpcId=vpc.vpc_id
+    )
+    sg.authorize_ingress(
+        IpProtocol="tcp",
+        FromPort=22,
+        ToPort=22,
+        CidrIp="0.0.0.0/0",
+    )
+    return sg
+
+
+@pytest.fixture(scope="module")
+def base_vars(subnets):
+    return {
+        "cluster_version": "1.27",
+        "name": str(uuid.uuid4()),
+        "subnet_ids": [subnets[0].id, subnets[1].id],
+    }
+
+
+@pytest.fixture(scope="module")
+def tf(this_dir: Path, moto_server) -> Terraform:
+    tf = Terraform(this_dir, f"http://localhost:{moto_server.port}")
+    tf.init(upgrade=True)
+    return tf
 
 
 @pytest.fixture(autouse=True)
@@ -12,8 +71,18 @@ def reset(moto_server: MotoServer, this_dir, tf) -> None:
     (this_dir / "terraform.tfstate").unlink(missing_ok=True)
 
 
+def test_defaults(eks_client: ..., base_vars: dict[str, Any], tf: Terraform):
+    tf.apply(vars=base_vars)
+    cluster = Cluster.from_name(eks_client, base_vars["name"])
+    assert cluster
+    assert cluster.version == base_vars["cluster_version"]
+    assert cluster.endpoint_private_access
+    assert cluster.endpoint_public_access
+    assert len(cluster.security_group_ids) == 0
+
+
 def test_name(eks_client: ..., tf: Terraform, base_vars: dict[str, Any]):
-    name = "my-custom-name"
+    name = str(uuid.uuid4())
     tf.apply(vars=base_vars | {"name": name})
     cluster = Cluster.from_name(eks_client, name)
     assert cluster.name == name
