--- test/irsa/conftest.py
@@ -1,17 +1,0 @@
-from pathlib import Path
-
-import pytest
-
-from ..utils import MotoServer, Terraform
-
-
-@pytest.fixture(scope="package")
-def this_dir() -> Path:
-    return Path(__file__).parent
-
-
-@pytest.fixture(scope="package")
-def tf(this_dir: Path, moto_server: MotoServer) -> Terraform:
-    tf = Terraform(this_dir, f"http://localhost:{moto_server.port}")
-    tf.init(upgrade=True)
-    return tf
